requirements.txt
tuya-connector-python
Flask
flask-socketio
python-dotenv